Attempt to move forward with Go 1.18
