+++ generate/maptoXXX.go.tmpl
@@ -4,7 +4,7 @@
 // This function streams continuously until the in-stream is closed at
 // which point the out-stream will be closed too.
 //
-// It is a special case of Map() aiming to avoid the use of functions C or SC for native Go types.
+// It is a special case of Map() aiming to avoid the use of functions CC or AC for native Go types.
 // It is provided as a convenience since Go 1.18 does not support parameterised methods.
 func (s Stream[T]) MapTo{{ .Name }}(mapper Function[T, {{ .GoType }}]) {{- if eq .Name "Int" "Int8" "Int16" "Uint32" "Uint64" "Uint" "Uint8" "Uint16" "Uint32" "Uint64" "Float32" "Float64" "String" }}Comparable{{ end }}Stream[{{ .GoType }}] {
 	st := NewConcurrentStream(orderlyConcurrentDo(s, mapper), s.concurrency)
@@ -21,7 +21,7 @@ func (s Stream[T]) MapTo{{ .Name }}(mapper Function[T, {{ .GoType }}]) {{- if eq
 // This function streams continuously until the in-stream is closed at
 // which point the out-stream will be closed too.
 //
-// It is a special case of FlatMap() aiming to avoid the use of functions C or SC for native Go types.
+// It is a special case of FlatMap() aiming to avoid the use of functions CC or AC for native Go types.
 // It is provided as a convenience since Go 1.18 does not support parameterised methods.
 func (s Stream[T]) FlatMapTo{{ .Name }}(mapper StreamFunction[T, {{ .GoType }}]) Stream[{{ .GoType }}] {
 	return NewConcurrentStream(orderlyConcurrentDoStream(s, mapper), s.concurrency)
